Restructured diagnostic endpoints Removed /diagnostic/block endpoint Moved /diagnostic storage and server endpoints to /node
