+++ rest/package.json
@@ -1,6 +1,6 @@
 {
   "name": "catapult-api-rest",
-  "version": "1.0.20.9",
+  "version": "1.0.20.10",
   "description": "",
   "main": "_build/index.js",
   "scripts": {
